Add exception handling to cuda/cublas executors
Apply will throw an exception if the launch of the
cuda/cublas function is invalid or fails.

Async will set the future exception state if an error
occurs during launch of the function.
